allow for frequency shift of less than a semitone

+++ src/melody.cpp
@@ -30,17 +30,17 @@ u32 Melody::get_total_n_samples(u32 sample_rate)
     return n_samples;
 }
 
-double up_semitones(double frequency, i16 semitones)
+double up_semitones(double frequency, double semitones)
 {
-    return frequency * exp2((double)semitones / 12.0f);
+    return frequency * exp2(semitones / 12.0f);
 }
 
-double up_octaves(double frequency, i16 octaves)
+double up_octaves(double frequency, double octaves)
 {
     return frequency * exp2(octaves);
 }
 
-float get_frequency(i16 octave, i16 semitones)
+float get_frequency(float octave, float semitones)
 {
     const double up_oct = up_octaves(C0_FREQUENCY, octave);
     const double up_semi = up_semitones(up_oct, semitones);
